Extract SpeedPills + SpeedFormatter to PracticeControls.swift (file length)
